Check if userId is present

+++ services/friendService.js
@@ -4,6 +4,10 @@ const friendRepository = require("../repositories/friendRepository");
 const acceptFriend = async (req, res) => {
     try {
         const { userId } = req.body;
+
+        if (!userId) {
+            return res.status(404).json({ "message": "Requête invalide" });
+        }
         await friendRepository.acceptFriend(userId, req.user.id);
         res.sendStatus(200);
     } catch (_) {
@@ -33,6 +37,10 @@ const sendFriendRequest = async (req, res) => {
 const deleteFriend = async (req, res) => {
     try {
         const { userId } = req.params;
+
+        if (!userId) {
+            return res.status(404).json({ "message": "Requête invalide" });
+        }
         await friendRepository.deleteFriend(req.user.id, userId);
         res.sendStatus(204);
     } catch (_) {
